WIP: Replace python-based JS extraction script with run_shell
This removes the python dependency in favor of a heavier reliance the
shell.

While not ideal right now, a resolution of
https://github.com/bazelbuild/bazel/issues/5511 would allow us to do
the file grepping in starlark

We could use entirely native rules this way, only calling out to the
`jar` binary directly, which should be included with any jdk, and
therefore result in a rule that is much leaner on dependencies while
also more portable.

Note: This hasn't been tested yet. I've thought a few times I did
using the JS example, but each time it has turned out bazel ended up
using different rules. This is a start, and an approach, but not a
ready patch.

This also still crucially lacks an implementation for .js.map files,
which is technically trivial, but I'd like to ideally do it without
invoking more shells.